style: enhance GraphQLExplorer UI and responsiveness
